+++ src/components/ui/category-icon/category-icon.tsx
