+++ src/components/ui/category-icon/index.ts
@@ -1,0 +1,1 @@
+export * from './category-icon'
