+++ src/components/ui/form/error-text/index.tsx
@@ -1,45 +1,1 @@
-/**
- * フォーム用ErrorTextコンポーネント
- *
- * バリデーションエラーメッセージを統一されたスタイルで表示するコンポーネントです。
- * エラーがない場合は何も表示されません。
- */
-
-import { type ComponentPropsWithRef } from 'react'
-import { tv, type VariantProps } from 'tailwind-variants'
-
-// スタイル定義
-const errorTextVariants = tv({
-  base: 'pt-1 text-sm font-bold text-error',
-  variants: {
-    size: {
-      sm: 'text-xs',
-      base: 'text-sm',
-    },
-  },
-  defaultVariants: {
-    size: 'base',
-  },
-})
-
-type CommonVariantProps = VariantProps<typeof errorTextVariants>
-
-type Props = ComponentPropsWithRef<'div'> & {
-  /** エラーメッセージ（空の場合は何も表示されない） */
-  error?: string
-  /** テスト用のdata-testid */
-  testId?: string
-} & CommonVariantProps
-
-export function ErrorText({ error, size, testId, ...props }: Props) {
-  // エラーがない場合は何も表示しない
-  if (!error) return null
-
-  const className = errorTextVariants({ size })
-
-  return (
-    <div {...props} data-testid={testId} className={className}>
-      {error}
-    </div>
-  )
-}
+export * from './error-text'
